Use models dev to get the limit, 32k otherwise

+++ pkg/modelsdev/types.go
@@ -46,8 +46,8 @@ type Cost struct {
 
 // Limit represents the context and output limitations of a model
 type Limit struct {
-	Context int `json:"context"`
-	Output  int `json:"output"`
+	Context int   `json:"context"`
+	Output  int64 `json:"output"`
 }
 
 // Modalities represents the supported input and output types
